ci: initial build workflow
This is an initial workflow to build the QLI 1.x (scarthgap) branch
along with meta-qcom-3rdparty.

This is using a specific persistent folder, to avoid mixing sstate and
downloads between QLI 1.x and QLI mainline.

This initial workflow builds all valid combinations based on the QLI
1.4 release notes. E.g.
 - DISTRO=qcom-wayland
 - QCOM_SELECTED_BSP=custom for all machines
 - QCOM_SELECTED_BSP=base for a subset of the machines
And only builds the qcom-console-image.

This initial workflow only builds the regular manifest. Realtime,
robotics or MM SDKs will be added later.

meta-qcom-3rdparty is added to the Yocto build configuration.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>