ci: move to IT managed AWS self hosted runners
IT provides a new set of AWS self hosted runners, both amd64 and arm64
with the following tags:

runs-on: [self-hosted, qcom-u2404, amd64-dev]
runs-on: [self-hosted, qcom-u2404, arm64-dev]

Note: '-dev' is a temporary tag to test a new type of runner with
local SSD.

Unlike with the previous builders, the container image is a vanilla
ubuntu image, this patch introduces changes to use the KAS container
image to run the build pipeline. Some adjustments are made to
accomodate this image.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>