ci: add a warmup build
The matrix jobs results in a large amount of jobs being started at the
same time. Running all jobs in parallel is not ideal to leverage the
OE sstate cache. For any common tasks (native, armv8, .. ) when the
object is missing from the sstate cache, it's likely going to be a
miss for several jobs. As a result, each of them will build the missed
object, and each of them will attempt to create the object in the
sstate cache.

Instead, with this change, we first build for the qcs8300 platform,
for both BSP variants. If these builds succeed we carry on with the other
machines/configs, with the hope that the sstate cache hit will be
better.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>